Move form to a different file
