+++ frontend/src/metabase/auth/components/ForgotPassword/ForgotPassword.styled.tsx
@@ -3,30 +3,6 @@ import { color } from "metabase/lib/colors";
 import Icon from "metabase/components/Icon";
 import Link from "metabase/core/components/Link";
 
-export const FormTitle = styled.div`
-  color: ${color("text-dark")};
-  font-size: 1.25rem;
-  font-weight: 700;
-  line-height: 1.5rem;
-  text-align: center;
-  margin-bottom: 1.5rem;
-`;
-
-export const FormFooter = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  margin-top: 1.5rem;
-`;
-
-export const FormLink = styled(Link)`
-  color: ${color("text-dark")};
-
-  &:hover {
-    color: ${color("brand")};
-  }
-`;
-
 export const InfoBody = styled.div`
   display: flex;
   flex-direction: column;
