+++ frontend/src/metabase/auth/components/ForgotPasswordForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./ForgotPasswordForm";
